requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# AI/ML Libraries
torch==2.1.0
transformers==4.35.0
langchain==0.0.335
openai==1.3.0
anthropic==0.7.1
google-generativeai==0.3.1
scikit-learn==1.3.2
networkx==3.2.1
gymnasium==0.29.1

# Data Processing
numpy==1.26.2
pandas==2.1.3
pydantic==2.5.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1

# Visualization & Frontend Support
jinja2==3.1.2
plotly==5.18.0
matplotlib==3.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
ipython==8.18.1
jupyter==1.0.0
streamlit==1.28.2

# Utilities
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
colorama==0.4.6
tqdm==4.66.1